[refactor] TiemItem 분리 및 상태 호이스팅 작업
